+++ engine/runtime-parser/src/main/java/org/enso/compiler/core/TreeToIr.java
@@ -243,20 +243,19 @@ final class TreeToIr {
         var methodRef = translateMethodReference(fn.getName(), false);
         var args = translateArgumentsDefinition(fn.getArgs());
         var body = translateExpression(fn.getBody());
-        var returnSignature = fn.getReturns();
-        if (returnSignature != null) {
-          var returnType = translateType(returnSignature.getType());
-          // TODO(#8240): Make use of return type declaration.
-        }
+        var loc = getIdentifiedLocation(inputAst, 0, 0, null);
+        var returnSignature = resolveReturnTypeSignature(fn);
         if (body == null) {
             var error = translateSyntaxError(inputAst, new Syntax.UnsupportedSyntax("Block without body"));
             yield join(error, appendTo);
         }
+
+        var ascribedBody = addTypeAscription(body, returnSignature, loc);
         var binding = new Method.Binding(
           methodRef,
           args,
-          body,
-          getIdentifiedLocation(inputAst, 0, 0, null),
+          ascribedBody,
+          loc,
           meta(), diag()
         );
         yield join(binding, appendTo);
@@ -396,7 +395,7 @@ final class TreeToIr {
         } else {
           name = buildNameOrQualifiedName(fun.getName());
         }
-        var ir = translateFunction(fun, name, fun.getArgs(), fun.getBody());
+        var ir = translateFunction(fun, name, fun.getArgs(), fun.getBody(), resolveReturnTypeSignature(fun));
         yield join(ir, appendTo);
       }
 
@@ -405,7 +404,7 @@ final class TreeToIr {
       case Tree.Assignment assignment -> {
         var name = buildName(assignment.getPattern());
         java.util.List<ArgumentDefinition> args = java.util.Collections.emptyList();
-        var ir = translateFunction(assignment, name, args, assignment.getExpr());
+        var ir = translateFunction(assignment, name, args, assignment.getExpr(), null);
         yield join(ir, appendTo);
       }
 
@@ -498,13 +497,15 @@ final class TreeToIr {
         return new Operator.Binary(fn, in, args.head(), getIdentifiedLocation(app), meta(), diag());
       }
     }
-    private Expression translateFunction(Tree fun, Name name, java.util.List<ArgumentDefinition> arguments, final Tree treeBody) {
+    private Expression translateFunction(Tree fun, Name name, java.util.List<ArgumentDefinition> arguments, final Tree treeBody, Expression returnType) {
       List<DefinitionArgument> args;
       try {
         args = translateArgumentsDefinition(arguments);
       } catch (SyntaxException ex) {
         return ex.toError();
       }
+
+      var loc = getIdentifiedLocation(fun);
       var body = translateExpression(treeBody);
       if (args.isEmpty()) {
         if (body instanceof Expression.Block block) {
@@ -522,18 +523,41 @@ final class TreeToIr {
         if (body == null) {
           body = translateSyntaxError(fun, Syntax.UnexpectedExpression$.MODULE$);
         }
-        return new Expression.Binding(name, body,
-          getIdentifiedLocation(fun), meta(), diag()
-        );
+
+        var ascribedBody = addTypeAscription(body, returnType, loc);
+        return new Expression.Binding(name, ascribedBody, loc, meta(), diag());
       } else {
         if (body == null) {
           return translateSyntaxError(fun, Syntax.UnexpectedDeclarationInType$.MODULE$);
         }
-        return new Function.Binding(name, args, body,
-          getIdentifiedLocation(fun), true, meta(), diag()
-        );
+
+        var ascribedBody = addTypeAscription(body, returnType, loc);
+        return new Function.Binding(name, args, ascribedBody, loc, true, meta(), diag());
       }
    }
+
+   /** Returns the return type of a function, if it was specified inline. */
+   private Expression resolveReturnTypeSignature(Tree.Function fun) {
+     var returnSignature = fun.getReturns();
+     if (returnSignature == null) {
+       return null;
+     }
+
+     return translateType(returnSignature.getType());
+   }
+
+  /**
+   * Wraps a body expression in a type ascription that will ensure that the type of the body is checked to match the provided type.
+   * <p>
+   * If the type is {@code null}, the body is returned unchanged.
+   */
+  private Expression addTypeAscription(Expression body, Expression type, Option<IdentifiedLocation> loc) {
+     if (type == null) {
+       return body;
+     }
+
+     return new Type.Ascription(body, type, loc, meta(), diag());
+  }
 
   private Type.Ascription translateTypeSignature(Tree sig, Tree type, Expression typeName) {
     var fn = translateType(type);
@@ -969,7 +993,7 @@ final class TreeToIr {
       }
       case Tree.Function fun -> {
         var name = buildName(fun.getName());
-        yield translateFunction(fun, name, fun.getArgs(), fun.getBody());
+        yield translateFunction(fun, name, fun.getArgs(), fun.getBody(), resolveReturnTypeSignature(fun));
       }
       case Tree.OprSectionBoundary bound -> translateExpression(bound.getAst(), false);
       case Tree.UnaryOprApp un when "-".equals(un.getOpr().codeRepr()) ->