+++ CHANGELOG.md
@@ -1006,6 +1006,7 @@
 - [HTTPS and WSS support in Language Server][7937]
 - [Export of non-existing symbols results in error][7960]
 - [Upgrade GraalVM to 23.1.0 JDK21][7991]
+- [Added opt-in type checks of return type][8502]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
 [3248]: https://github.com/enso-org/enso/pull/3248
@@ -1156,6 +1157,7 @@
 [7937]: https://github.com/enso-org/enso/pull/7937
 [7960]: https://github.com/enso-org/enso/pull/7960
 [7991]: https://github.com/enso-org/enso/pull/7991
+[8502]: https://github.com/enso-org/enso/pull/8502
 
 # Enso 2.0.0-alpha.18 (2021-10-12)